added the setting of the defaultFields in the correct place SSCS-10895
