+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/hmctsresponsereviewed/HmctsResponseReviewedAboutToSubmitHandler.java
@@ -55,7 +55,6 @@ public class HmctsResponseReviewedAboutToSubmitHandler extends ResponseEventsAbo
         setCaseCode(preSubmitCallbackResponse, callback);
         checkMandatoryFields(preSubmitCallbackResponse, sscsCaseData);
         setDwpDocuments(sscsCaseData);
-        setDefaultOverrideFields(sscsCaseData);
 
         if (isNull(sscsCaseData.getDwpResponseDate())) {
             sscsCaseData.setDwpResponseDate(LocalDate.now().toString());
@@ -122,18 +121,5 @@ public class HmctsResponseReviewedAboutToSubmitHandler extends ResponseEventsAbo
                 dwpDocument.getValue().setDocumentFileName(documentTypePrefix + " on " + todayDate);
             }
         }
-    }
-
-    private void setDefaultOverrideFields(SscsCaseData sscsCaseData) {
-        OverrideFields defaultOverrideFields = sscsCaseData.getSchedulingAndListingFields().getDefaultOverrideFields();
-        if (isNull(defaultOverrideFields)) {
-            defaultOverrideFields = OverrideFields.builder()
-                .poToAttend(YesNo.isYes(sscsCaseData.getDwpIsOfficerAttending()) ? YesNo.YES : YesNo.NO)
-                .build();
-        } else {
-            defaultOverrideFields.setPoToAttend(YesNo.isYes(sscsCaseData.getDwpIsOfficerAttending())
-                ? YesNo.YES : YesNo.NO);
-        }
-        sscsCaseData.getSchedulingAndListingFields().setDefaultOverrideFields(defaultOverrideFields);
     }
 }
