+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/updatelistingrequirements/UpdateListingRequirementsAboutToSubmitHandler.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.updatelistingrequirements;
 
+import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
 import static java.util.Objects.requireNonNull;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute.LIST_ASSIST;
@@ -15,8 +16,10 @@ import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingState;
+import uk.gov.hmcts.reform.sscs.ccd.domain.OverrideFields;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.State;
+import uk.gov.hmcts.reform.sscs.ccd.domain.YesNo;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.resendtogaps.ListAssistHearingMessageHelper;
 
@@ -51,6 +54,8 @@ public class UpdateListingRequirementsAboutToSubmitHandler implements PreSubmitC
 
         PreSubmitCallbackResponse<SscsCaseData> callbackResponse = new PreSubmitCallbackResponse<>(sscsCaseData);
 
+        setDefaultOverrideFields(sscsCaseData);
+
         State state = callback.getCaseDetails().getState();
         HearingRoute hearingRoute = sscsCaseData.getSchedulingAndListingFields().getHearingRoute();
 
@@ -78,5 +83,18 @@ public class UpdateListingRequirementsAboutToSubmitHandler implements PreSubmitC
             }
         }
         return callbackResponse;
+    }
+
+    private void setDefaultOverrideFields(SscsCaseData sscsCaseData) {
+        OverrideFields defaultOverrideFields = sscsCaseData.getSchedulingAndListingFields().getDefaultOverrideFields();
+        if (isNull(defaultOverrideFields)) {
+            defaultOverrideFields = OverrideFields.builder()
+                .poToAttend(YesNo.isYes(sscsCaseData.getDwpIsOfficerAttending()) ? YesNo.YES : YesNo.NO)
+                .build();
+        } else {
+            defaultOverrideFields.setPoToAttend(YesNo.isYes(sscsCaseData.getDwpIsOfficerAttending())
+                ? YesNo.YES : YesNo.NO);
+        }
+        sscsCaseData.getSchedulingAndListingFields().setDefaultOverrideFields(defaultOverrideFields);
     }
 }
