+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/hmctsresponsereviewed/HmctsResponseReviewedAboutToSubmitHandlerTest.java
@@ -56,7 +56,6 @@ public class HmctsResponseReviewedAboutToSubmitHandlerTest {
                     new DynamicListItem("reviewByTcw", "Review by TCW"), null))
                 .benefitCode("002")
                 .issueCode("CC")
-                .dwpIsOfficerAttending("Yes")
                 .build();
 
         sscsCaseDataBefore = SscsCaseData.builder().build();
@@ -469,30 +468,5 @@ public class HmctsResponseReviewedAboutToSubmitHandlerTest {
         assertThat(response.getWarnings().size(), is(0));
         assertEquals("Benefit code cannot be changed to the selected code",
             response.getErrors().stream().findFirst().get());
-    }
-
-    @Test
-    public void givenAHmctsResponseReviewedEventWithDwpIsOfficerAttending_thenSetPoToAttendValue() {
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(
-            ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-
-        assertNotNull(response.getData().getSchedulingAndListingFields()
-            .getDefaultOverrideFields().getPoToAttend());
-        assertEquals(YES, response.getData().getSchedulingAndListingFields()
-            .getDefaultOverrideFields().getPoToAttend());
-    }
-
-    @Test
-    public void givenDefaultOverrideValuesWithNoAttendWillBeUpdatedByDwpIsOfficerAttendingValue() {
-        OverrideFields defaultOverrideFields = OverrideFields.builder()
-            .poToAttend(NO)
-            .build();
-
-        callback.getCaseDetails().getCaseData().getSchedulingAndListingFields()
-            .setDefaultOverrideFields(defaultOverrideFields);
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(
-                ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-        assertEquals(YES, response.getData().getSchedulingAndListingFields()
-            .getDefaultOverrideFields().getPoToAttend());
     }
 }
